Move API wrapper from ui to common.
