+++ heron/ui/src/python/handlers/api/topology.py
@@ -26,7 +26,7 @@ import datetime
 
 import metrics
 from heron.ui.src.python.handlers import base
-from heron.ui.src.python.handlers import access
+from heron.common.src.python.handler import access
 from heron.ui.src.python.handlers import common
 
 class TopologyExceptionSummaryHandler(base.BaseHandler):
