+++ heron/ui/src/python/handlers/topology.py
@@ -24,7 +24,7 @@ import tornado.gen
 import urllib2
 from datetime import datetime
 
-import access
+from heron.common.src.python.handler import access
 import base
 import common
 from common.graph import TopologyDAG
